requirements.txt
# In terminal write pip install -r requirements.txt
pyserial
python-socketio[asyncio_client]
